Change resources copied to jar: no certificates
Change the pom configuration to override the default behavior of plugin
in charge of copying the resource folder to ther packaged jars. Now
only the application properties files will be copied, to get a
reference inside the jar, but the runtime file used for the properties
is the one outside the jar in the target folder.

The reason for the change is that there are security issues with
storing certificates inside jars that will be avoided this way. This
also fixes the issues with certificate path not being found, as before
the certificates could only be found inside the jar, and the paths
provided as input were referenced from the classpath (inside the jar).
While actually the path should start from the working directory, from
where the "java -jar" command was executed.

Also add new certificates for the workflow manager system. They use the
default cloud and root certificates as provided in the core systems
repository.