Update system startup and shutdown
Add root logger level, INFO, in static initialization block.

Update the check Core Systems method, to have a finite time of attempts
to contact the mandatory core systems. It is able to check the core
systems availability but does not work as intended in case some core
system is not ready at startup. In case core system is not present at
startup, the wanted behavior is to retry contacting attemps every 5 sec
for some minutes, given as input to method. The actual behavior is that
upon unsucessful communication, an exception is thrown in parallel
netty thread which stops and exits the system.

Update Arrowhead system creation and added a HttpJsonCloudPlugin to
facilitate Arrowhead cloud integration.

Update shutdown service, it now executes Arrowhead System shutdown, not
Java application shutdown. Although not sure if it unregisters the
services offered in the Service Registry