Fix method to test core systems at initialization
The checkCoreSystems method now works as expected, not only when the
first try of communication is succesful. When core system not present,
it will keep trying to contact each core system for a certain amount of
minutes, sending a new request every 5 seconds.

Problem was with exception not being caught in HttpClientRequest. The
exception thrown, ConnectException, is not known at compile time as it
comes inside a Future. It is not thrown by any method directly called
in my code, but from the netty library executing the send request in
another thread. Therefore I was not catching it, only the ones I knew
of.

Solution involved the use of methods provided to handle Futures in
Kalix library, and the creation of a new private class to handle the
time duration of these requests. This class was created to include all
logic inside the lambda functions used in the methods for Futures, as
lambda functions can only access static and instance variables or final
local variables.