Add support for product authentication
Add internal variable to Workflow Manager to keep a set of valid
products that are expected to order some operation by sending a POST
request to the system. Now the list of products only comes from the
properties file but in the future a Factory system will provide this
service.