Add POST workflow service and fix code format
Add code to POST endpoint of workflow service. Before it was just used
to test that endpoint could be reached, now it includes logic to test
that the consumer is an authorized product and then it will resolve the
request by calling the WExecutor which offer the requested workflow.
The service is still incomplete, error handling has to be improved.

The code was reformat to improved visibility by reducing the maximum
lenght of the code lines.