Restructure GET /operations and fix bug compare WExecutor
Logic in GET /operations service in now wrapped in
WManagerMain#updateWorkflow(). This will support to call the service
logic from other services without needing to consume the service as an
external system.

Found bug in WManagerMain#executeWorkflow(). When WExecutor system
stored internally as providing workflow is shutdown, and it tries to
find such service, it will remove all the workflows provided by that
WExecutor system from internal memory. This removal is done by
comparing the system that used to provide the workflow, with all other
workflows stored. Before the comparison will always fail.