Add DTO for POST /operations response and improve endpoint
Create new DTO class to be received by product when it requests an
operation. This DTO serves as a ticket; it includes a unique
operationID as a new variable added to WManagerMain, an optional
productID given when the config option to test products is true, and
some data given by theWorkflow Executor when a new workflow starts
execution.

As there were some issues when testing the endpoint, the test1 variable
was added until the issue is resolved. This endpoint is not fully
functional yet.