Fix POST /operations thread block and improve endpoint
There was an issue that seemed like a Future never ending. This was
caused by an extra call to await() that blocked the thread managing the
request. This issue is fixed now.

Also fix issues with the response not being updated inside the
processing of the request. This was due to how Futures behave, as their
execution was done at a later time after the service response was
built and sent. This was solved by returning the Future of the request,
now the service waits for Futures to finish and specify the response.

Found another issue with DTO parsing. Now the service will discard the
input as is not a valid DTO, not sure why the method bodyAs() always
returns with DtoReadException.