Fix possible problems with thread blocking and test logs
Remove calls to await() inside route handlers as they could block the
thread and stall the system. Used on*() instead.

The problem with DTOs was not in the code, but in the input. I was
sending a list of objects, and inpuct expects just an object.

Also tested to add logger for Kalix library only, not netty and find
problems easier.