Change DTO naming and fix POST /operations
Change name of DTO interfaces to be more meaningful. Also tested the
possibility of modifying the equals() method of a DTO interface, but it
is not possible. Instead we can use a default method to implement the
functionality wanted, but this will not work for objects part of data
structures, as they call equals() and hasCode().

Fix POST /operations service/endpoint. It had a comparison between two
Workflow DTOs that was throwing errors, as one of the fields,
workflowConfig was optional and changed. The solution was to split the
map workflowToExecutor in two fields, one map to store the relation
workflow name to WExecutor system, and a set to store the original
workflow types as provided by the WExecutor system. The map is useful
to find the right WExecutor system while the set is useful when
querying the operations of the workstation.

Also reformatted code. Split the method executeWorkflow into smaller
methods as the same procedures are call in two program flows.
Cut lines to not go further than line maximum length.