Add classes representing Arrowhead systems and Product
Split code in WManagerMain as class was getting too big. The part
related with checking core systems and the time count was moved to a
new package, that can be reused for other Arrowhead projects as is.

Inside this new package I created classes to represent an Arrowhead
system and its child class a Smart Product. These classes are used in
WManager to know which products are present in the Workstation and
make the relation between the WorkflowId, coming from Workflow Executor,
the operationId, coming from Workflow Manager, and the Smart Product
which is acted upon.

Also added how to create/obtain a Smart Product from a service request
and how to build the StartOperationDto.