Add initial structure of workflow result service
Add new service provided for Workflow Executor System to provide the
results as the Http initial request would close after certain time.

Tested the interaction of this new service with Workflow Executor and
of the echo service. The echo service went through a change in encodings
to be compatible with system created using arrowhead client library as
they accept "text/plain" and "application/json" but kalix library only
works with "application/*" encodings, so the requests were not accepted.

The logic to store and resend the workflow results is still missing.